requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
django-filter==24.2
django-guardian==2.4.0
django-cors-headers==4.3.1
Markdown==3.6
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1